server/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
llama-cpp-python>=0.1.62
python-multipart>=0.0.6
